Revert "Refactoring navigation"
