+++ src/App.tsx
@@ -1,12 +1,33 @@
 import * as React from "react";
-import { Route } from "react-router-dom";
+import { Redirect, Route, Switch } from "react-router-dom";
+import { DashboardTabBar } from "./components";
+import { TabBar } from "./components/TabBar.component";
 import "./index.css";
-import { TabBarPage } from "./TabBarPage";
 
 export default () => {
+  const renderDefault = ({ match }: { match: any }) => {
+    return (
+      <React.Fragment>
+        <DashboardTabBar />
+        <Redirect to={"/china/dc/summary_current_forecast"} />
+      </React.Fragment>
+    );
+  };
+  const renderPage = ({ match }: { match: any }) => {
+    return (
+      <React.Fragment>
+        <DashboardTabBar />
+      </React.Fragment>
+    );
+  };
+
   return (
     <main>
-      <Route path="/:department" component={TabBarPage} />
+      <TabBar />
+      <Switch>
+        <Route exact={true} path="/" render={renderDefault} />
+        <Route path="/:page/:category" render={renderPage} />
+      </Switch>
     </main>
   );
 };
