--- src/pages/Demand/KPIS/index.ts
@@ -1,5 +1,0 @@
-export { KPISChinaDC } from "./KPIS_ChinaDC.component";
-export { KPISChinaDI } from "./KPIS_ChinaDI.component";
-export { KPISInternationalDI } from "./KPIS_InternationalDI.component";
-export { KPISInternationalEIB } from "./KPIS_InternationalEIB.component";
-export { KPISInternationalIL } from "./KPIS_InternationalIL.component";
