--- src/pages/Demand/PreviousForecasts/PreviousForecasts_ChinaDC.component.tsx
@@ -1,48 +1,0 @@
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
-import Grid from "@material-ui/core/Grid";
-import { withStyles } from "@material-ui/core/styles";
-import Typography from "@material-ui/core/Typography";
-import * as React from "react";
-
-const styles = {
-  root: {
-    padding: 24
-  },
-  font: {
-    fontWeight: 700,
-    fontSize: 16
-  }
-};
-const PreviousForeCastsComponent = (props: any) => {
-  const { classes } = props;
-
-  return (
-    <div className={classes.root}>
-      <Grid container={true} spacing={24}>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Previous Offtake
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Previous Offtake
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-      </Grid>
-    </div>
-  );
-};
-
-export const PreviousForecastsChinaDC = withStyles(styles)(
-  PreviousForeCastsComponent
-);
