--- src/pages/Demand/PreviousForecasts/index.ts
@@ -1,15 +1,0 @@
-export {
-  PreviousForecastsChinaDC
-} from "./PreviousForecasts_ChinaDC.component";
-export {
-  PreviousForecastsChinaDI
-} from "./PreviousForecasts_ChinaDI.component";
-export {
-  PreviousForecastsInternationalDI
-} from "./PreviousForecasts_InternationalDI.component";
-export {
-  PreviousForecastsInternationalEIB
-} from "./PreviousForecasts_InternationalEIB.component";
-export {
-  PreviousForecastsInternationalIL
-} from "./PreviousForecasts_InternationalIL.component";
