--- src/pages/Demand/SummaryCurrentForecast/SummaryCurrentForecast_ChinaDC.component.tsx
@@ -1,102 +1,0 @@
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
-import Grid from "@material-ui/core/Grid";
-import { withStyles } from "@material-ui/core/styles";
-import Typography from "@material-ui/core/Typography";
-import * as React from "react";
-import { LineColumnChart } from "src/charts/LineColumnChart";
-import { WaterfallChart } from "src/charts/WaterfallChart";
-import { CustomTable } from "src/table/Table";
-import { CustomTable as TableNut } from "src/table2/Table";
-
-const styles = {
-  root: {
-    padding: 24
-  },
-  font: {
-    fontWeight: 700,
-    fontSize: 16
-  }
-};
-const SummaryCurrentForecastComponent = (props: any) => {
-  const { classes } = props;
-
-  return (
-    <div className={classes.root}>
-      <Grid container={true} spacing={24}>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Current Cycle Phasing - DC offtake Quarterly Phasing
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Current cycle DC
-              </Typography>
-            </CardContent>
-            <CustomTable />
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Cycle evolution - DC total offtake bridge by brand (2018,2019)
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Risks & opportunities
-              </Typography>
-            </CardContent>
-            <WaterfallChart />
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trade flow DC
-              </Typography>
-            </CardContent>
-            <LineColumnChart />
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trade Flow DC by brand - NC
-              </Typography>
-            </CardContent>
-            <TableNut />
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trade Flow DC by brand - AP
-              </Typography>
-            </CardContent>
-            <TableNut />
-          </Card>
-        </Grid>
-      </Grid>
-    </div>
-  );
-};
-
-export const SummaryCurrentForecastChinaDC = withStyles(styles)(
-  SummaryCurrentForecastComponent
-);
