--- src/pages/Demand/SummaryCurrentForecast/SummaryCurrentForecast_InternationalEIB.component.tsx
@@ -1,113 +1,0 @@
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
-import Grid from "@material-ui/core/Grid";
-import { withStyles } from "@material-ui/core/styles";
-import Typography from "@material-ui/core/Typography";
-import * as React from "react";
-import { CustomTable } from "src/table/Table";
-
-const styles = {
-  root: {
-    padding: 24
-  },
-  font: {
-    fontWeight: 700,
-    fontSize: 16
-  }
-};
-const SummaryCurrentForecastComponent = (props: any) => {
-  const { classes } = props;
-
-  return (
-    <div className={classes.root}>
-      <Grid container={true} spacing={24}>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Current Cycle Phasing - EIB offtake Quarterly Phasing
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trackable offtake
-              </Typography>
-            </CardContent>
-            <CustomTable />
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Non-trackable offtake
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Total offtake
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trade flow - EIB (total)
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Trade Flow
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Risks & opportunites - 2018 EIB Offtake (Oct cycle - Nov cycle)
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Risks & opportunites - 2019 EIB Offtake (Oct cycle - Nov cycle)
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-        <Grid item={true} xs={12}>
-          <Card>
-            <CardContent>
-              <Typography gutterBottom={true} className={classes.font}>
-                Bridge between cycles
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-      </Grid>
-    </div>
-  );
-};
-
-export const SummaryCurrentForecastInternationalEIB = withStyles(styles)(
-  SummaryCurrentForecastComponent
-);
