--- src/pages/Demand/SummaryCurrentForecast/index.ts
@@ -1,15 +1,0 @@
-export {
-  SummaryCurrentForecastChinaDC
-} from "./SummaryCurrentForecast_ChinaDC.component";
-export {
-  SummaryCurrentForecastChinaDI
-} from "./SummaryCurrentForecast_ChinaDI.component";
-export {
-  SummaryCurrentForecastInternationalDI
-} from "./SummaryCurrentForecast_InternationalDI.component";
-export {
-  SummaryCurrentForecastInternationalIL
-} from "./SummaryCurrentForecast_InternationalIL.component";
-export {
-  SummaryCurrentForecastInternationalEIB
-} from "./SummaryCurrentForecast_InternationalEIB.component";
