--- src/pages/Demand/index.ts
@@ -1,6 +1,0 @@
-export {
-  SummaryDefinition
-} from "./SummaryDefinition/SummaryDefinition.component";
-export {
-  SummaryKeyAssumption
-} from "./SummaryKeyAssumption/SummaryKeyAssumption.component";
