+++ src/pages/SummaryDefinition/SummaryDefinition.component.tsx
@@ -4,7 +4,7 @@ import Grid from "@material-ui/core/Grid";
 import { withStyles } from "@material-ui/core/styles";
 import Typography from "@material-ui/core/Typography";
 import * as React from "react";
-import { colors } from "../../../theme";
+import { colors } from "../../theme";
 
 const styles = {
   root: {
