+++ src/pages/index.ts
@@ -1,0 +1,6 @@
+export {
+  SummaryDefinition
+} from "./SummaryDefinition/SummaryDefinition.component";
+export {
+  SummaryKeyAssumption
+} from "./SummaryKeyAssumption/SummaryKeyAssumption.component";
